Swap v3 with v4
