+++ versione3/Package/__init__.py
@@ -1,2 +1,2 @@
 __all__ = ['analyseBot', 'screenBot',
-           'solverBot', 'decoratori', 'print_coloured' , 'webhook' , 'game_controller' , 'globals']
+           'solverBot', 'decoratori', 'print_coloured']
