+++ versione4/Package/__pycache__/globals.cpython-311.pyc
