+++ versione4/Package/globals.py
