+++ versione4/__pycache__/test.cpython-312.pyc
