+++ versione4/__pycache__/test2.cpython-312.pyc
